Round-17 policy fixes: one major type per file
The remaining companion types (ghost add button, transparent overlay,
focus field, pressed-dim constant, content container, flipped document
view) move into their own files.
