+++ Sources/Sidebar/AppKitList/Cells/SidebarRowChecklistFlippedView.swift
@@ -1,0 +1,8 @@
+import AppKit
+import CmuxWorkspaces
+import SwiftUI
+
+@MainActor
+final class SidebarRowChecklistFlippedView: NSView {
+    override var isFlipped: Bool { true }
+}
