+++ Sources/Sidebar/AppKitList/Cells/SidebarRowPressedDim.swift
@@ -1,0 +1,7 @@
+import CoreGraphics
+
+/// The SwiftUI plain-button pressed treatment the legacy checklist/status
+/// Buttons had: content dims to half strength while the mouse is down.
+enum SidebarRowPressedDim {
+    static let pressedAlpha: CGFloat = 0.5
+}
