+++ Sources/Sidebar/AppKitList/Cells/SidebarRowTaskStatusGlyphButton.swift
@@ -169,9 +169,3 @@ final class SidebarRowTaskStatusGlyphButton: NSControl {
         return true
     }
 }
-
-/// The SwiftUI plain-button pressed treatment the legacy checklist/status
-/// Buttons had: content dims to half strength while the mouse is down.
-enum SidebarRowPressedDim {
-    static let pressedAlpha: CGFloat = 0.5
-}
