+++ Sources/Sidebar/AppKitList/Cells/SidebarWorkspaceRowCellView.swift
@@ -1272,13 +1272,6 @@ final class SidebarWorkspaceRowTableCellView: NSTableCellView {
     }
 }
 
-/// Flipped host for the row's content subviews; its `alphaValue` implements
-/// the legacy Done-row dim (content composites as one group at 60%).
-@MainActor
-final class SidebarRowContentContainerView: NSView {
-    override var isFlipped: Bool { true }
-}
-
 /// Borderless single-line rename field (select-all on focus, Escape cancels).
 @MainActor
 final class SidebarRowInlineRenameField: NSTextField, NSTextFieldDelegate {
